requirements.txt
polygon-api-client
matplot
pandas
numpy<2
scipy==1.10.1
pmdarima
mysql-connector-python
python-dotenv
streamlit